+++ src/Onebrb.Data/IOnebrbContext.cs
@@ -6,7 +6,7 @@ namespace Onebrb.Data
 {
     public interface IOnebrbContext
     {
-        Task<Item> GetItemAsync(long itemId);
+        Task<Item> GetItemAsync(int itemId);
         Task<ICollection<Item>> GetItemsAsync(string userId);
     }
 }
