+++ src/Onebrb.Services/Services/IItemService.cs
@@ -7,7 +7,7 @@ namespace Onebrb.Services.Services
 {
     public interface IItemService
     {
-        Task<ItemServiceModel> GetItemAsync(long itemId);
+        Task<ItemServiceModel> GetItemAsync(int itemId);
         Task<ICollection<ItemServiceModel>> GetItemsAsync(string userId);
     }
 }
